Always run Pivot Table queries immediately
